creating integration tests for group metadata templates
