+++ config/locales/en.yml
@@ -92,6 +92,7 @@ en:
         user_id: User
       metadata_template:
         description: Description
+        fields: Metadata fields
         name: Name
       namespaces/project_namespace:
         description: Description
